backendmane/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
openai==1.12.0
twilio==8.10.0
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.16.3
requests==2.31.0
pydantic==2.5.3
pdfplumber==0.10.3
